+++ packages/framework/src/Concerns/HasPageMetadata.php
@@ -19,7 +19,7 @@ trait HasPageMetadata
 
     public function getCanonicalUrl(): string
     {
-        return Hyde::url(Hyde::pageLink($this->getCurrentPagePath().'.html'));
+        return Hyde::url(Hyde::formatHtmlPath($this->getCurrentPagePath().'.html'));
     }
 
     public function getDynamicMetadata(): array